Fix segfault caused by appending 'type' to string

+++ include/argsparser/parser.h
@@ -62,7 +62,7 @@ namespace ArgsParser
                 return container->getConvertedValue();
             };
 
-            const int& error_code;
+            const unsigned short& error_code;
             const std::string& error_description;
 
             /**
